social-scraper/requirements.txt
yt-dlp>=2024.1.0
instaloader>=4.10
pandas>=2.0.0
requests>=2.31.0
aiohttp>=3.9.0
tqdm>=4.66.0
python-dotenv>=1.0.0
